requirements.txt
pandas>=2.0.0
biopython>=1.81
multiqc>=1.15
matplotlib>=3.7.0
seaborn>=0.12.0
numpy>=1.24.0
